test: Fix AssemblyAnalyzerTest to be robust to Grok availability (#7950)

+++ core/src/main/java/org/owasp/dependencycheck/analyzer/AssemblyAnalyzer.java
@@ -427,7 +427,9 @@ public class AssemblyAnalyzer extends AbstractFileTypeAnalyzer {
      */
     @Override
     public void closeAnalyzer() throws Exception {
-        FileUtils.delete(grokAssembly.getParentFile());
+        if (grokAssembly != null) {
+            FileUtils.delete(grokAssembly.getParentFile());
+        }
     }
 
     @Override
